Remove Devise from sign up / conf controllers
**Why**: Hard to understand controllers when they inherit from an
external lib but also override it in many places.
